fix dto get restaurant export

+++ src/orders/orders.service.ts
@@ -42,7 +42,9 @@ export class OrdersService {
         order.orderDetails.forEach(async (orderDetail) => {
             await this.orderDetailService.delete(orderDetail.id);
         });
-        await this.paymentOrderService.delete(order.paymentOrder.id);
+        if (order?.paymentOrder) {
+            await this.paymentOrderService.delete(order.paymentOrder.id);
+        }
 
         return await order.remove();
     }
